+++ fsl/installer/fslinstaller.py
@@ -68,7 +68,7 @@ log = logging.getLogger(__name__)
 __absfile__ = op.abspath(__file__).rstrip('c')
 
 
-__version__ = '3.4.1'
+__version__ = '3.4.2'
 """Installer script version number. This must be updated
 whenever a new version of the installer script is released.
 """
@@ -76,6 +76,10 @@ whenever a new version of the installer script is released.
 
 DEFAULT_INSTALLATION_DIRECTORY = op.join(op.expanduser('~'), 'fsl')
 """Default FSL installation directory. """
+
+
+DEFAULT_ROOT_INSTALLATION_DIRECTORY = '/usr/local/fsl/'
+"""Default FSL installation directory when the installer is run as root. """
 
 
 FSL_RELEASE_MANIFEST = 'https://fsl.fmrib.ox.ac.uk/fsldownloads/' \
@@ -1218,6 +1222,9 @@ class Context(object):
         if self.__destdir is not None:
             return self.__destdir
 
+        if os.getuid() != 0: defdestdir = DEFAULT_INSTALLATION_DIRECTORY
+        else:                defdestdir = DEFAULT_ROOT_INSTALLATION_DIRECTORY
+
         # The loop below validates the destination directory
         # both when specified at commmand line or
         # interactively.  In either case, if invalid, the
@@ -1232,13 +1239,13 @@ class Context(object):
                 printmsg('\nWhere do you want to install FSL?',
                          IMPORTANT, EMPHASIS)
                 printmsg('Press enter to install to the default location '
-                         '[{}]\n'.format(DEFAULT_INSTALLATION_DIRECTORY), INFO)
+                         '[{}]\n'.format(defdestdir), INFO)
                 response = prompt('FSL installation directory [{}]:'.format(
-                    DEFAULT_INSTALLATION_DIRECTORY), QUESTION, EMPHASIS)
+                    defdestdir), QUESTION, EMPHASIS)
                 response = response.rstrip(op.sep)
 
                 if response == '':
-                    response = DEFAULT_INSTALLATION_DIRECTORY
+                    response = defdestdir
 
             response  = op.expanduser(op.expandvars(response))
             response  = op.abspath(response)
@@ -2016,6 +2023,9 @@ def parse_args(argv=None, include=None):
                   of None for all arguments that are not included.
     """
 
+    if os.getuid() != 0: destdir = DEFAULT_INSTALLATION_DIRECTORY
+    else:                destdir = DEFAULT_ROOT_INSTALLATION_DIRECTORY
+
     username = os.environ.get('FSLCONDA_USERNAME', None)
     password = os.environ.get('FSLCONDA_PASSWORD', None)
 
@@ -2044,6 +2054,7 @@ def parse_args(argv=None, include=None):
         'miniconda'       : (None, {}),
         'no_self_update'  : (None, {'action'  : 'store_true'}),
         'exclude_package' : (None, {'action'  : 'append'}),
+        'root_env'        : (None, {'action'  : 'store_true'}),
     }
 
     if include is None:
@@ -2053,11 +2064,13 @@ def parse_args(argv=None, include=None):
         'version'      : 'Print installer version number and exit',
         'listversions' : 'List available FSL versions and exit',
         'dest'         : 'Install FSL into this folder (default: '
-                         '{})'.format(DEFAULT_INSTALLATION_DIRECTORY),
+                         '{})'.format(destdir),
         'overwrite'    : 'Delete existing destination directory if it exists, '
                          'without asking',
         'no_env'       : 'Do not modify your shell or MATLAB configuration '
-                         'implies --no_shell and --no_matlab)',
+                         '(implies --no_shell and --no_matlab). When running '
+                         'the installer script as the root user, the root '
+                         'shell profile is never modified.',
         'no_shell'     : 'Do not modify your shell configuration',
         'no_matlab'    : 'Do not modify your MATLAB configuration',
         'fslversion'   : 'Install this specific version of FSL',
@@ -2119,6 +2132,11 @@ def parse_args(argv=None, include=None):
         # fnmatch-style wildcard pattern. Can
         # be used multiple times.
         'exclude_package' : argparse.SUPPRESS,
+
+        # If the installer is run as root, the
+        # --no_env flag is automatically enabled
+        # UNLESS this flag is also provided.
+        'root_env'        : argparse.SUPPRESS,
     }
 
     # parse args
@@ -2147,6 +2165,11 @@ def parse_args(argv=None, include=None):
             printmsg('Home directory {} does not exist!'.format(args.homedir),
                      ERROR, EMPHASIS)
             sys.exit(1)
+
+    # --no-env is automatically enabled
+    #  when installer is run as root
+    if os.getuid() == 0 and not (args.root_env):
+        args.no_env = True
 
     # don't modify shell profile
     if args.no_env: